# contracts/api/v2/openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "additionalProperties": false,
        "properties": {
          "download_allowed": {
            "description": "Whether the effective policy permits downloads",
            "examples": [
              true
            ],
            "type": "boolean"
          },
          "email": {
            "description": "Contact email; empty when none is set",
            "examples": [
              "alice@example.test"
            ],
            "type": "string"
          },
          "id": {
            "description": "Account identifier",
            "examples": [
              "1"
            ],
            "minLength": 1,
            "type": "string"
          },
          "impersonation": {
            "$ref": "#/components/schemas/Impersonation",
            "description": "Present only while an administrator impersonates this account"
          },
          "permissions": {
            "description": "Effective assignable permissions; empty for a disabled account",
            "examples": [
              [
                "marker_edit"
              ]
            ],
            "items": {
              "description": "Assignable permission name",
              "examples": [
                "marker_edit"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "role": {
            "description": "Server-wide role",
            "enum": [
              "admin",
              "user"
            ],
            "examples": [
              "user"
            ],
            "type": "string"
          },
          "username": {
            "description": "Login name",
            "examples": [
              "alice"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "permissions",
          "download_allowed"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "Safe occurrence-specific explanation; not for control flow",
            "examples": [
              "The request did not pass validation; see errors."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Field-level validation details",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "type": "array"
          },
          "instance": {
            "description": "urn:silo:request:\u003crequest-id\u003e, matching the X-Request-ID response header",
            "examples": [
              "urn:silo:request:000000000000000000000003"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, equal to the response status",
            "examples": [
              422
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "Short summary fixed for the problem type",
            "examples": [
              "Validation failed"
            ],
            "type": "string"
          },
          "type": {
            "description": "Stable problem type URI; the final segment is the problem identifier",
            "examples": [
              "https://siloserver.org/docs/api/v2/problems/validation_failed"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "type": "object"
      },
...
      "SetupStatus": {
        "additionalProperties": false,
        "properties": {
          "needs_setup": {
            "description": "True until the first administrator account exists",
            "examples": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "needs_setup"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A session token or an API key in the Authorization header. The server decides which it holds.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Silo API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/account/me": {
      "get": {
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get the authenticated caller's login account.",
        "tags": [
          "account"
        ],
        "x-silo-class": "authenticated",
        "x-silo-service-backed": true
      }
    },
...
    "/api/v2/auth/impersonation/end": {
      "post": {
        "operationId": "endImpersonation",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Return an impersonating session to the administrator who started it.",
        "tags": [
          "auth"
        ],
        "x-silo-class": "authenticated",
        "x-silo-service-backed": true
      }
    },
...
    "/api/v2/auth/sessions": {
      "get": {
        "operationId": "listSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSessionCollection"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List the caller's login sessions, active and revoked.",
        "tags": [
          "auth"
        ],
        "x-silo-class": "authenticated",
        "x-silo-service-backed": true
      }
    },
...
    "/api/v2/system/setup": {
      "get": {
        "operationId": "getSetupStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Report whether the server still needs its first administrator.",
        "tags": [
          "system"
        ],
        "x-silo-class": "public",
        "x-silo-service-backed": true
      }
    },
...
  }
}